% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Req_data}
\alias{Req_data}
\title{data_inputs}
\usage{
Req_data(
  industry = req_industry,
  active = req_active,
  num_employee = req_num_employee,
  creation_date = req_date_creation
)
}
\arguments{
\item{industry}{The type of industry a business is classified under, according to the Registraire des entreprises}

\item{active}{a dummy variable equal to 1 if a company exists currently}

\item{num_employee}{a range of employees employed by the business}

\item{creation_date}{date the business was created (all else equal, all businesses before a specified date will be called)}
}
\value{
All corresponding businesses in Quebec
}
\description{
This function takes the data inputs of industry, an active dummy , the number of employees, and the creation date and returns the corresponding businesses
If any/all inputs are left blank, all possible observations are called
}
\examples{
Req_data()
}
