% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Req_employee}
\alias{Req_employee}
\title{employee_query}
\usage{
Req_employee()
}
\value{
table displaying the ranges of employees in businesses and the corresponding codes
}
\description{
This function provides a table displaying the range of employees that corresponds to the number of employees code, and vice-versa
}
\examples{
Req_employee()
}
