% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Req_industry}
\alias{Req_industry}
\title{industry_query}
\usage{
Req_industry(industry)
}
\arguments{
\item{industry}{The name of the industry (If a full name isn't specified, all industries with a matching string will be called)}
}
\value{
The corresponding industry code
}
\description{
This function provides an industry code to the corresponding inputted industry number
}
